ruler: gracefully stop subservices when starting fails (#6414)
### Context

Starting can fail for two reasons: 1) the process was sent SIGINT, 2) an actual error occurred while starting up; for exmaple the KV store became unavailable

In those cases the ruler service's starting function may return an error. This automatically transitions the BasicService into a Failed state. This means that the sotpping function is not ran. This in turn means that the subserices such as the lifecycler aren't stopped. In other words, the ruler doesn't remove itself from the ring.

This PR explicitly shuts down subservices when the startup failed to make sure at least the lifecycler is properly stopped.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>