Operations: Use metric_name as trigger name
In KEDA 2.10 the metricName was deprecated and is now auto-generated.

We use the metric_name in our dashboard mixins to disambiguate which
metric is being used in the scaling.

Instead place the metric_name as the name for the trigger. This becomes
available as a `scaler` label on the keda_* metrics.