ingester_test.go: update exemplar related unit tests
The tests make it seem like you need two write requests to ingest
an exemplar: one to create the series with a sample and one to
append the exemplar. Actually a single write request is enough, but
the content needs to be in order of sample and then exemplar.

Two new test cases added:
- single write request where there is a timeseries that has both a
sample and an exemplar in the same timeseries.
- single write request where there are two timeseries with the first
having a sample and the second the exemplar.

Currently Prometheus and Grafana Agent use the second form to send
exemplars, which is a bit wasteful.

Also added native histogram version of tests.

Noticed when fixing: https://github.com/prometheus/prometheus/pull/13113

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>