querier: track consistency check failures without retries (#6590)
* querier: track consistency check failures without retries

Consistency checks are currently tracked per attempt. For example, this means that if there were two attempts to fetch some blocks and the first was partially siccessful and the second fetched the missing blocks, then the ratio of failed checks will be 50%. This is misleading because the second check succeeded and the query overall was successful. We graph this ratio in Mimir / Queries.

### Note to reviewers

with this change I took the chance to refactor the consistency checker. One side effect of this change is that the upload and deletion marker grace periods will not be recalculated upon each retry. This means that if a request to the store-gateway takes long enough for a block to fall outside of the uploadGracePeriod, then that block won't be queried. The time between query-ingesters-within and query-stores after is 1h and upload-grace-period is 45m. So unless a request takes 15m, this shouldn't be a problem (assuming default values).

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Add CHANGELOG.md entry

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Track checksTotal along with checksFailed

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Properly size queried map

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Sort blocks in newStoreConsistencyCheckFailedError

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Update pkg/querier/blocks_consistency_checker.go

Co-authored-by: Charles Korn <charleskorn@users.noreply.github.com>

* Update info on cortex_querier_blocks_consistency_checks_total

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Update description on "Consistency checks failed" panels

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>
Co-authored-by: Charles Korn <charleskorn@users.noreply.github.com>