Speed up some store-gateway tests (#6671)
* storegateway tests: disable tsdb transaction isolation

Mimir always disables it, so no point in running tests with it enabled.
This helps tests to run faster.

* storegateway tests: commit appender periodically

Change the function to get an appender factory so we can commit
after a while and get a new appender. This avoids building up massive
arrays of data inside the appender, and helps to speed up tests.

* storegateway tests: reduce overhad from testify

Testify methods like `assert.Equalf` call `t.Helper()` which examines
the current call stack; this is very expensive to do in the inner loop.
Modify the calls to avoid `t.Helper()` in the happy path.

* storegateway tests: re-use details when creating test block

It's much faster to call the appender interface re-using the labels
and series reference returned.