querier: backoff retrying to send responses to query-frontend
### Background

The querier will try sending a query's response to a frontend. If the sending fails, then the querier will remove the connection from the pool and immediately try to establish a new one.

### Problem

When there are multiple queries, then the can remove each other's new connections which leads to a live lock until the retries expire.
One reason why the sending can fail is that the client was cleaned up

### What this PR does

* make retrying less aggressive to make livelocks less likely, but also to give time to whatever was causing to error to recover
* don't remove the query-frontend connection from the pool; if the connection is indeed faulty, then the client pool will clean it up (within 6 seconds of it "becoming" faulty)

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>