Increase default timeout for index-queries cache requests
As use of the index-queries cache has increased in the store-gateway, we
see more and more timeouts making requests to it. Timeouts making a request
mean the connection is closed and reopened since it assumed to be "bad". This
can cause a huge number of connections on nodes to end up in `TIME_WAIT`.

Increase the timeout used by the index-queries cache to 450ms from the default
of 200ms in jsonnet and helm.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>