mixin: use heatmap for per pod p99 latency
I was doing an investigation and looking at "Per pod p99 latency" panels when there are many pods becomes really noisy and difficult to tell whether only one pod has high latency or all of them.

I'm proposing this change of per-pod latency panels to heatmaps to help with this.

The tradeoff is that exemplaers are now harder to navigate because they are all on the same scale.

I'll leave this open for some time for more people to chime in.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>