Relaxed the hash ring heartbeat period and timeout for distributor, ingester, store-gateway and compactor.
These are values which help reduce the pressure on a KV store or reduce the CPU spent by memberlist in passing messages.
The tradeoff is that abrupt shutdowns/crashes of components will take longer to detect by peers.
We've been running with these values at Grafana Labs for some time and haven't seen problems.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>
Signed-off-by: Marco Pracucci <marco@pracucci.com>