Fix max_fetched_series_per_query not applied to /series from ingesters
in queriers.

The code is slightly inefficient as it checks for unique hashes twice,
once in distributor.go and once in query_limiter.go. On the other hand the
limiter uses a lock, so it makes sense to first merge the results from
ingesters and throw out the duplicates before calling the limiter under lock.

I chose pkg/distributor and not pkg/querier for the placement to align
with the limit check in the ingester streaming code in pkg/distributor/query.go

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>