querier: don't log error on stream connection reset
If gRPC server configured with connection_max_age, it closes the
client's connections after the max_age + age_grace period.

At the moment there is nothing we expect a user to act after
the errors, the client stream logs when the connection was reset.
We move the logging to the debug level, to reduce any possible
confusion.