Extensible "Rollout progress" workload grouping
We have some cells where we have multiple deployments (10+) for the same
workload. They make the rollout progress dashboard unreadable.

In this PR I'm extracting the `label_replace(...)` calls from the
"Rollout progress" panel into an array in the config, so users can
extend that with their workload matchers.

I also used this opportunity to:
- Replace `cortex_service` label with `workload`.
- Change query from regexp-matching on the metric name to
  kube_statefulset_... or kube_deployment_...: this is required for
  Adaptive Metrics to correctly process this metric name lookup.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>