Reset compat gauges each 30 minutes elapsed
This commit resets the compat gauges in Alertmanager each 30 minutes
that has elapsed. Due to how config.Load works in Alertmanager,
it is not possible to reset the gauge per tenant each time a
configuration is reloaded in Mimir. In order to ensure old metrics
about incompatible configurations are aged out over time, we instead
reset the gauges using a 30 minute timer. This is not ideal, but it
is a short-lived measure until we can remove the
--utf8-strict-mode-enabled flag in a few months time. Delete this
code when removing the flag.