store-gateway: remove cortex_bucket_store_blocks_loaded_by_duration
The PR which added this metric 6074, the motivation was to help detect compactors which are falling behind. There is another metric `cortex_bucket_store_series_blocks_queried` added by jhalterman in 7112 which serves better the purpose of detecting uncompacted blocks. The reason is that it has an explicit label for uncompacted blocks, and it also is more sensitive to recent blocks (assuming those are queried much more often than old uncompacted blocks).

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>