Deregister compat metrics
This commit deregisters the compat metrics in Mimir. These metrics
are not proving as useful as expected, due to the fact that the
compat package is global and Mimir reloads configurations at fixed
intervals. Instead, we use the logs from the compat package to
learn about incompatible inputs and disagreement.