Set series cache TTL different based on block metadata
Instead of using a default TTL of 7d for all index-cache related entries,
pick different TTLs based on metadata about the block. This allows entries
related to blocks that will be deleted (OOO blocks, 2h, 12h) to expire and
be eligible to be evicted sooner than they would otherwise.

This change only sets different TTLs for series entries in the cache since
they account for most of the space in the cache.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>