Use thread safe chunk iterator in sharding tests
The PromQL engine currently uses pooling of histogram pointers, optimized
for reading chunks where the iterator returns a copy of the data.
However in the sharding tests we run multiple engines on the same data
with an iterator from promql.StorageSeries that does not copy the histograms.
So it can happen that pointers are reused between the goroutines and
end up in a race.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>