Reduces dependencies of integration/ruler_test.go
Do not start querier as that's built into the ruler.
Do not try to query block storage as that might be slower than the
first test case query execution, resulting in the block storage
index load being context canceled. The cancellation would be cached
as well, leading to hard to debug problems.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>