Allow tenants to be included or excluded from store-gateways
Add configuration to the store-gateway to allow specific tenants to be
enabled or disabled using the same logic as the corresponding configuration
in the compactor. Useful for ensuring that a particular tenant generating
a huge number of blocks doesn't consume all disk space on store-gateways
while waiting for compaction to reduce block count.

Fixes #7031

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>