Make distributor's pushReq.Cleanup() calls safer
We follow a simple logic: if we don't call next(), we should call
pushReq.Cleanup(). However, there's nothing in the code ensuring that,
which leads to bugs and fixes like:

- https://github.com/grafana/mimir/pull/7176
- https://github.com/grafana/mimir/pull/7755

A simple way to prevent that is to couple the `cleanupInDefer = false`
to the `next()` call, and this does exactly that.

Note that `nextOrCleanup` is introduced just for convenience, and it's
inlined.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>