autoscaling: add `autoscaling_querier_predictive_scaling_enabled`
This PR upstreams a change that we have been trialing at GL.

### Problem

Some workloads have a very uneven distribution of queries. Some of the causes of this can be external alerting systems and reporting tools which run batch jobs once a day.

This results in an influx of queries at fairly regular intervals (~24h). Autoscaling can only respond so fast to incoming queries. In practice, it takes anywhere between 2 and 7 minutes to respond to queries and trigger a scaling event. As a result Mimir's query queues fill up and Mimir start returning HTTP 429 to new queries until queriers have had the time to scale out and accommodate the added load.

### Solution

Scale preemptively. Automated workloads are also highly predictive. This PR uses the value of the inflight queries at the same time last week (6d23h30m instead of 7d) to prepare for periodic queries.

### Results

We've seen HTTP 429 responses decrease significantly on a weekly basis: between 66% and 95%

### Side effects

There are a lot of false-positive scaling events. As a result of this we've seen querier cost increase between 20% and 50%. Given that querier cost is usually 1-2% of the cell's TCO, this is somewhat acceptable for a better QoS.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>