Allow decoding metrics requests twice
In a downstream implementation a middleware peeks into the request by
calling the `PrometheusCodec.DecodeMetricsQueryRquest()` method, and
sometimes ignores and just passes the request to the next handler.

However, the call to `r.Form()` has left the body consumed and the body
can't be decoded again.

Fixed that and added a test to make sure this doesn't happen again.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>