Mixin: Lower alerting threshold for BucketIndexNotUpdated alert
This change will cause us to page when a bucket index has not been updated
by two compactor cleanup cycles (every 15 minutes) with a 5 minute buffer to
avoid false-positives.

The update time of the index is checked when queriers perform queries so as
soon as it exceeds the 1 hour max age, queries begin to fail. Thus we need
to alert on the age of the index before it begins to fail queries.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>