Helm: Change Ruler to not use alertmanager headless service.
The jsonnet deployment configures the Ruler to use the cluster IP alertmanager
service, whereas the Helm deployment uses the headless service. This change
updates the Helm chart to match the jsonnet.

The reason we do not need to use a headless service for Mimir Alertmanager is
because it has a distributor built in, so it's only necessary to send alerts
to a single Alertmanager replica, and they will be then distributed to the
correct three replicas for the tenant.