Remove UTF-8 migration code
This commit removes the migration code added to the multi-tenant
Alertmanager as part of enabling UTF-8 strict mode in Grafana Cloud.
It was used to find incompatible Alertmanager tenant configurations
that would prevent UTF-8 strict mode from being enabled. However,
this is no longer needed as we have been running this mode for over
a month.