Improve logging of incompatible configurations
This commit improves the logging of tenant configurations that
are incompatible with UTF-8 to make it easier to understand
for Mimir operators.

It uses the UTF-8 matchers parser instead of the fallback parser
so we get errors instead of logs about invalid configurations.
While we still log the error at present, this will allow us to
set a gauge to 0 or 1 for each tenant.