Fix metadata ingestion test flake
This change fixes a metadata ingestion test flake caused by metadata being
ingested when it should have been rejected. The conversion of global to local
limits only counts the number of ingesters with tokens when determining how
to divide the global limit. When there are no ingesters with tokens due to
race conditions, 0 is used for the global limit which causes 0 to be used for
the local limit which is interpreted as "unlimited".

Fixes #7979

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>