Do not encode bucket count in the labels
Labels are already packed due to tag==stringlabels in TSDB so Add()
doesn't work on them. And after they are decoded for the response we'd
have to insert the bucket count label into the correctly sorted position
which would mean never being able to just send the encoded labels.
Instead of adding a new interface I'll add a backwards compatible array
of integers as a compromise.

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>