Fix flaky TestOwnedSeriesIngesterRingStrategyRingChanged
The test case `change of state is not interesting` changes the state of an ingester in the ring. It then expects that the change is classified as "not interesting. "However, it changes the state by adding the ingester again into the ring. This also changes the RegisteredTimestamp of the instance in the ring. Timestamps are unix timestamps in seconds. So if   `change of state is not interesting` runs in a different unix second than `new instance added`, then the test will detect a change in the RegisteredTimestamp, which is an "interesting" change and the test will fail.

> NB: The test assumes that its test cases run together. So each subtest requires the previous to have run. Otherwise, they all fail on their own.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>