store-gateway: Add timeout for index-header loading gate
This change introduces an optional timeout on the gate used to limit the
concurrency of index-header loads. This helps in cases where a store-gateway
may have to load a large index-header before it can serve a query. This
prevents and unbounded number of requests from blocking while index-headers
are loaded.

Fixes #8137

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>