Introduce global and per-tenant flags to control rule evaluation concurrency
This change introduces two flags that control if rule evaluation concurrently.

First, you have a flag that controls the total amount of rules that can be running concurrently at any given time per ruler replica with:

`-ruler.max-global-rule-evaluation-concurrency`

Then, this is paired with `-ruler.max-concurrent-rule-evaluations-per-tenant` to control the amount of rules a single tenant is allowed to have concurrently. By default, this is `4`. However, the behaviour is disabled by default because `-ruler.max-global-rule-evaluation-concurrency` is set to `0` by default.

Signed-off-by: gotjosh <josue.abreu@gmail.com>