Alertmanager: Implement experimental extended /receivers API.
Grafana Alertmanager has an extended version of the /receivers API that we would
like to provide compatibility for. This API is however not in upstream
Prometheus Alertmanager, and so the proposal is to add this API strictly under
an experimental banner, and not change the existing /receivers API to maintain
strict compatability with upstream.