Update regexp library to sort alternates
This brings in https://github.com/grafana/regexp/pull/11, so that long
lists of alternates that escape FastRegexpMatcher will get sorted and
hence execute faster.

The bulk of this PR is https://github.com/grafana/regexp/pull/10 which
brings the library up-to-date with Go 1.22.1; mostly cosmetic but a
couple of code improvements.