Extend max rules per rule group limit to be supported per namespace
Introduces a new limit named `ruler_max_rules_per_rule_group_by_namespace` which allows one to specify which namespaces do we want to be different to have a different limit in terms of rules per group.

With this change, we now have a hierarchy for this limit as follows:

1. Per tenant limit for the given namespace.
2. Default limit for the given namespace.
3. Per tenant limit set by RulerMaxRulesPerRuleGroup
4. Default limit set by RulerMaxRulesPerRuleGroup

This is part 1 of the change and I need to extend the same level of support to the other limit related to rule groups of `MaxRuleGroupsPerTenant`.

Signed-off-by: gotjosh <josue.abreu@gmail.com>