ingest storage: propagate only sampled traces
We've discovered internally at GL with another user of franz-go that the tracing info can be a significant overhead for small records.

This shouldn't be the case for ingest storage, but to be on the safe side, I'm proposing this change now that we know it may be a problem.

While testing this internally I couldn't spot a significant change in the number of bytes transferred.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>