Allow namespaces to be protected and deny modification unless specified
This is a draft of for allowing mimir to have a concept of protected namespaces. There are two key design pieces:

1. The use of `X-Override-Protection` as a header to override the protection and allow modification e.g. `X-Override-Protection: mynamespace,myothernamespace`
2. The use of `X-Protected-Namespaces` as a header from Mimir itself to tell the client that this namespace is actually protected whilst it's being read e.g. `X-Protected-Namespaces: mynamespace,myothernamespace`

It's important to note that both headers value can be a command-separeted lists of strings. In the case of `X-Protected-Namespaces` is to list _all_ the protected namespaces on a list request and on the case of `X-Override-Protection` is to include all namespaces this client is allowed to modify.

Marked as a draft as I'd like to get some early feedback on the design. There's stil plenty to do including tests.

Signed-off-by: gotjosh <josue.abreu@gmail.com>