2.13 release notes
This is a draft PR with release notes. I've based this hugely on the changelog. The wording in the release notes is also largely influenced by the changelog as well.

There are a lot of enhancements and bug fixes in 2.13, so it was difficult to include everything. I tried to focus on user-visible changes, changes that are likely to impact more Mimir installations, or changes related to new features.

I'm open to feedback for mentioning other features or removing some of the ones mentioned.

related to XXX

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>