[Cardinality API] Include missing labels in label values
When querying for cardinality by some label, the
result may not add up to the total cardinality.
This is because some series may be missing that
label entirely. Instead, this change will include
the empty value in the result. As an optimization,
the empty value is skipped if it would be filtered
out by matchers.