Refactor atomicfs package to expose file-like API
The current atomicfs package exposes methods that operate on paths and
write the entire contents of a file buffering it in memory first. This
means that they cannot operate on large files.

This change exposes an os.File like API so that the contents of a file
can be written without needing to keep the entire file in memory. It also
keeps a `CreateFile` method for simplicity when the amount of data being
written is small.

This also changes the behavior of the atomicfs package so that files
are _always_ renamed from a temporary file to the final path. The only
place in Mimir using atomicfs that did not already make use of renaming
these files is the shutdown marker functionality.

Related #8485

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>