2.13 release notes (#8445)
* 2.13 release notes

This is a draft PR with release notes. I've based this hugely on the changelog. The wording in the release notes is also largely influenced by the changelog as well.

There are a lot of enhancements and bug fixes in 2.13, so it was difficult to include everything. I tried to focus on user-visible changes, changes that are likely to impact more Mimir installations, or changes related to new features.

I'm open to feedback for mentioning other features or removing some of the ones mentioned.

related to XXX

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Update docs/sources/mimir/release-notes/v2.13.md

* Update docs/sources/mimir/release-notes/v2.13.md

Co-authored-by: Charles Korn <charleskorn@users.noreply.github.com>

* Update docs/sources/mimir/release-notes/v2.13.md

Co-authored-by: Charles Korn <charleskorn@users.noreply.github.com>

* Update docs/sources/mimir/release-notes/v2.13.md

Co-authored-by: Charles Korn <charleskorn@users.noreply.github.com>

* Update docs/sources/mimir/release-notes/v2.13.md

* Update wording on owned series

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Order feature highlights and remove space

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Apply styling suggestions

Co-authored-by: taylor c <41653732+tacole02@users.noreply.github.com>

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>
Co-authored-by: Charles Korn <charleskorn@users.noreply.github.com>
Co-authored-by: taylor c <41653732+tacole02@users.noreply.github.com>