Chore: replace store_query_gate_ismyturn span with a log
There's no need for this verbose span that is almost always just a few
microseconds log. We can just log on the parent span how much time we
spent waiting.

I was also tempted to avoid logging at all if it was <1ms wait,
but finally preferred to be consistent.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>