Validate `-ruler.independent-rule-evaluation-concurrency-min-duration-percentage`
To ensure that its value is 0 or greater.

NB: I double checked that the code doesn't panic with 0 and I'll add a test in https://github.com/grafana/mimir/pull/8846.

Do let me know if you think a unit test is also a good idea because in my head it seemd superflous.

Signed-off-by: gotjosh <josue.abreu@gmail.com>