store-gateway: fix flaky TestBucketStore_PersistsLazyLoadedBlocks
This test used to use `time.Sleep()` which is flaky. I changed it to use assert.Eventually(), which is less flaky (but still has the potential :()

I'm debating whether we should use fsnotify, but that adds an extra dependency (it's currently only an indirect dep), and we'll be leaking the name of the lazy-loaded.json outside the `indexheader` package. If reviewers think it's a good idea, I can make the change.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>