Code cleanup for `pusherConsumer`
This small refactor cleans up a bit the code within pusherConsumer to address some dependency inversion of some of the components.

Some of the things I've done are:

- Move functions around to comply with Go code's standards e.g. Functions that are part of the public interface go after the struct definition
- Avoid passing individual metrics around and instead have a single metrics struct that is passed around to the different structs
- Simplify the `consumerFactory` definition  to push the construction of the `pusherCloser` to itself.

There are other things I want to do such as: Remove the "multitenant pusher" abstraction but I think this is a enough for a first PR.

NB: This does not the change the behaviour of the code at all.
Signed-off-by: gotjosh <josue.abreu@gmail.com>