ingester: set active series default to 20m (#8975)
* ingester: set active series default to 20m

In GL we bill based on active series in the last 20 minutes. This is also the value we want bill GEM for. It would be easier to reason about active series if they have the same default value everywhere. Because of this I'm changing the default value from 10m to 20m.

I'll leave this open until the end of the week to gather more feedback.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Update CHANGELOG.md entry

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>

* Update CHANGELOG.md

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>
(cherry picked from commit 1501b90e1cd93e4c39c21fba9ded7987089f2512)