store-gateway: gracefully shut down BucketStores
BucketStores would get shut down immediately when we receive SIGTERM instead of waiting for ongoing bucket syncs to finish first. This sometimes results in leaving index header readers open before shutting down.

This made some tests flaky. It could possibly mean that `BucketStores` stop `BucketStore` instances before waiting for ongoing queries to finish, but I am not certain of this and couldn't find proof in 2 weeks worth of store-gateway and querier logs.

Still this is a bug and we should fix it.