Add a feature flag to enable logging when parsing label matchers
To validate UTF-8 strict mode is working as intended, Grafana Mimir
logs parsing label matchers to validate that the correct parser is
used and also identify any label matchers that are incompatible
or ambigious. However, this also produces a huge amount of log volume,
and doesn't need to be enabled all the time. This puts it behind
an experimental feature flag which will be removed when we also remove
the feature flag for UTF-8 strict mode.