kafka replay speed: address post-merge feedback from 9152 (#9202)
The suggestion in the PR was to make `concurrentFetchersMinBytesMaxWaitTime` a field on `concurrentFetchers`. But those are created and started by the `PartitionReader` in its `starting` function [here](https://github.com/grafana/mimir/blob/b5050f31e51b8b2224ae10983f5221c9368897e8/pkg/storage/ingest/reader.go#L196). So the tests don't have control over them.

This is another approach where we have the field on `PartitionReader` and pass it to the constructor of `concurrentFetchers`.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>