kafka replay speed: introduce kafka flavours (#9204)
Normally all kafka flavours support the same API, except when they don't. Warpstream treats the `MaxBytes` field of a Fetch request as a limit on the uncompressed bytes of all the records. Apache Kafka treats it as a limit on the compressed bytes.

This difference in behaviour is quite significant and leads to very suboptimal performance when using the wrong interpretation.

This PR adds a config flag to be able to choose between the Kafka flavours. Open to feedback on what to call the flag and what to call the options.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>