kafka replay speed: upstream partition_offset_reader.go (#9448)
This is the first of series of PRs to upstream the code for improving Kafka replay speed in the ingester.

In this PR I'm upstreaming a tiny change related to partitionOffsetReader. We need caching in the reader so that we can check the start offset of the partition. We don't need that to be very exact because we use it to find out if we're trying to consume from before the start.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>