kafka replay speed: rename existing cortex_ingest_storage_reader_records_* metrics
Because of the batching behaviour of the sharding pushers, there is no longer a 1:1 matching between the kafka records and the errors we get from the TSDB. Because of this, the existing metrics `cortex_ingest_storage_reader_records_failed_total` and `cortex_ingest_storage_reader_records_total` no longer make sense.

This PR renames them to `cortex_ingest_storage_reader_requests_total` and `cortex_ingest_storage_reader_requests_failed_total` while keeping the alerts and dashboards working. It also fixes a bug in how `requests_total` is tracked - it used to be only tracked when there is an error.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>