kafka replay speed: dont send requests for higher than the high watermark
This is related to a problem we have: high e2e latency when fetching at the head of the partition. The current theory is that WS introduces extra delay when requesting an offset which doesn't exist.

Currently, we have two sets of flags: one for startup and one for ongoing fetching. This PR enables us to have just a single configuration which determines the max concurrency. We will limit concurrency by never requesting an offset which doesn't exist.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>