Allow querying against an expandable number of storages
Currently the querier expects to run queries against two storages only:
  - Ingesters
  - Store gateways

The queriers know when each queryable was applicable and queries the
appropriate queryables.

This commmit introduces the TimeRangeQueryable struct which is a wrapper
around Queryable. It ties the Queryable to the logic that deems it
applicable to an incoming query.

This, in conjunction with the newly-added AdditionalStorageQueryables
property on the Mimir struct, enables the querier to query optionally
against more storages than just the ingesters and store gateways in the
future.