Upstream rollout-operator zone-tracker settings. (#9699)
Configures rollout-operator to use zone-tracker mode, which uses a configmap to store metadata about statefulset scaling operations rather than as attributes on the statefulset being operated on. We found that the latter caused infinite recursion-type behavior in the K8S clusters of several cloud service providers.

Part of this configuration is granting rollout-operator RBAC permissions to manipulate configmaps.