Don't free buffers after reading query stream
Follow up on https://github.com/grafana/mimir/pull/9401

While the write path has been carefully crafted to make sure that we
don't keep references to the strings backed by arrays with data from
grpc stream (because it would cause a memory leak), this doesn't seem to
be the case in the query path, where some tests started failing after we
started to reuse those backing arrays through the usage of memory
buffers implemented in the new gRPC library.

This change reverts the buffer freeing, means that it won't be recycled
(and will be garbage collected) to ensure data correctness, while we
investigate where the data references are kept.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>