chore: Remove flaky TestRulerEvaluationDelay test
The test is unreliable and we've been unable to figure out why it
sometimes fails. This leads to people blindly re-running failing CI
instead of assuming it has caught a legitimate issue. Remove the test
since it reduces confidence in the test suite.

Fixes https://github.com/grafana/mimir/issues/4857