Allow HTTP configuration for GCS and Azure storage clients
This change allows the HTTP client used by GCS and Azure object storage
clients to be configured in the same way that we allow the S3 object
storage client to be configured. This will allow us to increase the
number of idle connections kept around in busy cells and avoid connection
churn when performing many object storage requests at once.

Fixes #9777