ingester: shortcut fetchWants that have already been satisfied
After 9855 there were some flaky tests (`TestConcurrentFetchers/consume_from_end_and_update_immediately` in particular) because we'd buffer records and then discard them at PollFetches.

This PR shortcuts this, so the tests are less likely to be flaky. Now we can still make a request for a fetchWants that's been satisfied, but it's less likely and there will only be up to `concurrency` of these.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>